refactor: Split update UI models into dedicated files
- Moved `AvailableUpdateUiData` into its own file
- Created new files for `UpdatesSettingsUiData` and `UpdatesInfoUiData`
- Updated `UpdatesUiData` to reference new UI data classes
- Adjusted imports in `UpdatesViewModel`
- Removed inline `AvailableUpdateUiData` from `DownloadsUiData`
