+++ app/src/main/kotlin/org/strigate/ferrot/presentation/model/AvailableUpdateUiData.kt
@@ -1,0 +1,6 @@
+package org.strigate.ferrot.presentation.model
+
+data class AvailableUpdateUiData(
+    val tag: String,
+    val localFilePath: String?,
+)
