+++ app/src/main/kotlin/org/strigate/ferrot/presentation/model/DownloadsUiData.kt
@@ -4,8 +4,3 @@ data class DownloadsUiData(
     val downloads: List<DownloadItemUiData>,
     val availableUpdate: AvailableUpdateUiData?,
 )
-
-data class AvailableUpdateUiData(
-    val tag: String,
-    val localFilePath: String?,
-)
