+++ app/src/main/kotlin/org/strigate/ferrot/presentation/model/UpdatesInfoUiData.kt
@@ -1,0 +1,6 @@
+package org.strigate.ferrot.presentation.model
+
+data class UpdatesInfoUiData(
+    val lastAvailableUpdateCheckMillis: Long,
+    val lastDependencyUpdateCheckMillis: Long,
+)
