+++ app/src/main/kotlin/org/strigate/ferrot/presentation/model/UpdatesSettingsUiData.kt
@@ -1,0 +1,6 @@
+package org.strigate.ferrot.presentation.model
+
+data class UpdatesSettingsUiData(
+    val automaticUpdates: Boolean,
+    val automaticDependencyUpdates: Boolean,
+)
