+++ app/src/main/kotlin/org/strigate/ferrot/presentation/viewmodel/UpdatesViewModel.kt
@@ -17,8 +17,8 @@ import org.strigate.ferrot.analytics.AnalyticsLogger
 import org.strigate.ferrot.domain.usecase.SettingsUseCase
 import org.strigate.ferrot.domain.usecase.StateUseCase
 import org.strigate.ferrot.extensions.toast
-import org.strigate.ferrot.presentation.model.UpdatesInfo
-import org.strigate.ferrot.presentation.model.UpdatesSettings
+import org.strigate.ferrot.presentation.model.UpdatesInfoUiData
+import org.strigate.ferrot.presentation.model.UpdatesSettingsUiData
 import org.strigate.ferrot.presentation.model.UpdatesUiData
 import org.strigate.ferrot.presentation.state.UpdatesUiState
 import org.strigate.ferrot.work.DownloadAvailableUpdateWorker
@@ -47,11 +47,11 @@ class UpdatesViewModel @Inject constructor(
         ) { automaticUpdates, automaticDependencyUpdates, lastAvailableUpdateCheckMillis, lastDependencyUpdateCheckMillis ->
             UpdatesUiState.Data(
                 UpdatesUiData(
-                    settings = UpdatesSettings(
+                    settings = UpdatesSettingsUiData(
                         automaticUpdates = automaticUpdates,
                         automaticDependencyUpdates = automaticDependencyUpdates,
                     ),
-                    info = UpdatesInfo(
+                    info = UpdatesInfoUiData(
                         lastAvailableUpdateCheckMillis = lastAvailableUpdateCheckMillis,
                         lastDependencyUpdateCheckMillis = lastDependencyUpdateCheckMillis,
                     ),
